Aggregate splice variants

Merge all splice variants of an expression dataset into unique locus, unaware of alternative splicing, by summing all variants of the same gene

aggregate_splice_variants(data)

Arguments

data
expression dataframe with splice variants as rownames and samples as columns

Value

dataframe with aggregated rows

Examples

data("abiotic_stresses")
head(aggregate_splice_variants(abiotic_stresses$normalized_counts))

#>                 C_1       C_2       C_3       S_1      S_2       S_3       M_1
#> AT1G01010 126.67770  68.11023  65.42961  93.86575  87.5764  94.94861  65.05674
#> AT1G01020 207.37239 221.48363 186.59926 192.22508 223.7622 195.83770 234.04437
#> AT1G01030  32.61701  34.50651  55.74004  33.55201  15.8055  30.98948  21.38578
#> AT1G01040 857.11927 981.83053 987.53763 896.31808 832.9201 939.18927 948.16950
#> AT1G01050 844.65099 842.80140 836.30033 834.90591 813.9337 784.34086 801.46705
#> AT1G01060 656.52962 700.96360 635.81601 531.93922 408.5573 514.54424 741.10719
#>                 M_2       M_3       H_1       H_2       H_3      SM_1      SM_2
#> AT1G01010 100.40712 126.60277  95.94273 135.09481 117.88109  95.26083 103.22907
#> AT1G01020 197.11029 179.12095 314.68009 366.25704 319.84802 166.48480 177.58564
#> AT1G01030  29.83183  33.44789  47.97137  39.02739  51.39696  24.82495  30.85007
#> AT1G01040 904.16462 988.55996 802.84153 701.59237 908.14710 881.28582 870.42671
#> AT1G01050 762.21318 718.78024 927.44640 969.98073 936.20924 853.99808 796.16902
#> AT1G01060 783.03539 876.33479  36.90879  20.81461  20.61913 844.83649 843.43290
#>                SM_3      SH_1      SH_2      SH_3      MH_1      MH_2      MH_3
#> AT1G01010  96.93197  95.62140  96.96100 101.98612  83.07132 107.83420  97.72878
#> AT1G01020 184.27943 344.11588 355.75908 347.33674 332.28529 317.07673 329.62076
#> AT1G01030  33.99041  61.59351  66.08684  61.61452  45.14746  36.34635  56.26199
#> AT1G01040 827.23183 790.21445 783.35607 702.72766 846.06337 813.47552 782.73606
#> AT1G01050 830.09730 903.91001 790.21700 855.25379 692.26103 705.64132 722.44818
#> AT1G01060 867.94128  29.78703  39.95479  19.22937  33.81043  40.06131  16.80814
#>               SMH_1     SMH_2     SMH_3
#> AT1G01010 100.03786  96.92257 122.39208
#> AT1G01020 267.53381 314.42203 320.85215
#> AT1G01030  45.97144  42.79828  57.07611
#> AT1G01040 806.69890 787.80910 857.84991
#> AT1G01050 835.28114 971.53100 841.26970
#> AT1G01060  48.96958  33.77757  26.12639