Expression matrix from Arabidopsis thaliana.

Expression matrix from Arabidopsis thaliana.

demo_data_At

Format

A list with the following elements:

raw_counts

raw transcript aboundances as obtained after mapping and quantifying RNASeq reads (dataframe of integer positive values).

conditions

Experimental conditions of each sample, regardless of biological replicates. (character vector)

design

For each condition, the level of the corresponding factors.

Source

BPMP Institute, Supagro Montpellier

Examples

{
 head(demo_data_At$raw_counts)
 demo_data_At$conditions
 demo_data_At$design
}

#>     Nitrate Iron
#> cnf       1    1
#> cnF       1    0
#> cNF       0    0
#> cNf       0    1