Targets and regulators of a gene in a network

Targets and regulators of a gene in a network

describe_node(graph, node)

Arguments

graph

igraph object

node

gene to describe

Value

list with targets and regulators attributes

Examples

data("abiotic_stresses")
data("regulators_per_organism") 
mat <- abiotic_stresses$heat_DEGs_regulatory_links
network <- DIANE::network_thresholding(mat, n_edges = 600)
data <- network_data(network, regulators_per_organism[["Arabidopsis thaliana"]])
nodes <- data$nodes[order(-data$nodes$degree),]
gene_of_interest <- nodes$id[1] 
DIANE::describe_node(network, node =gene_of_interest)
#> $targets
#>  [1] "AT1G44830" "AT3G23230" "AT2G41070" "AT1G75040" "AT2G43590" "AT1G72920"
#>  [7] "AT4G11470" "AT1G21120" "AT1G35230" "AT4G18253" "AT3G57240" "AT1G60750"
#> [13] "AT1G05675" "AT5G25910" "AT1G16090" "AT4G19975" "AT4G37400" "AT4G39610"
#> [19] "AT2G28560" "AT4G25070" "AT5G01740" "AT1G15125" "AT5G39670" "AT4G18250"
#> [25] "AT5G57220" "AT2G37980" "AT1G80660" "AT4G29050" "AT3G53250" "AT3G60730"
#> [31] "AT3G20395" "AT4G18205" "AT4G07995" "AT1G75750" "AT3G01960" "AT5G54710"
#> [37] "AT2G21840" "AT3G23510" "AT1G47480" "AT1G08300" "AT1G44970" "AT5G44565"
#> [43] "AT4G19430" "AT5G41590" "AT4G34770" "ATCG00170" "AT5G54720" "AT4G31354"
#> [49] "AT4G13320" "AT1G72645" "AT2G35850" "AT1G71000" "ATMG00080" "AT5G60310"
#> [55] "AT4G29880" "AT4G35160" "AT5G12020" "AT3G27640" "AT1G64561" "AT4G06534"
#> [61] "AT4G06536" "AT1G66500" "AT4G29770" "AT3G28220" "AT3G57460" "AT2G29300"
#> [67] "AT4G15780" "AT4G16590" "AT5G42020" "ATMG00090" "AT3G23550" "ATMG00650"
#> [73] "AT5G47960" "AT1G30070" "AT4G37022" "AT1G14360" "AT5G58770" "AT4G30570"
#> [79] "AT1G52560" "AT1G07400" "AT2G40390" "AT4G15550" "AT5G22520" "AT1G03470"
#> [85] "AT5G22785" "AT3G27550" "ATMG00180" "AT5G02570" "AT2G26860" "AT5G23660"
#> [91] "AT5G66053"
#> 
#> $regulators
#> character(0)
#>