Plots the histogram of in and out degrees, betweeness of regultors and target genes.

Plots the histogram of in and out degrees, betweeness of regultors and target genes.

draw_network_degrees(nodes, graph)

Arguments

nodes
dataframe containing the nodes information

graph
igraph object

Examples

data("demo_data_At")
data("regulators_per_organism")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)

#> TCC::INFO: Calculating normalization factors using tmm ...

#> TCC::INFO: Done.

threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

#> Warning: norm factors don't multiply to 1

topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes

regressors <- intersect(genes, regulators_per_organism[["Arabidopsis thaliana"]])
mat <- DIANE::network_inference(normalized_counts, conds = demo_data_At$conditions,
targets = genes, regressors = regressors)
network <- DIANE::network_thresholding(mat, n_edges = length(genes))
DIANE::draw_network_degrees(data$nodes, network)

#> Error in data$nodes: object of type 'closure' is not subsettable