Enriched ontologies for recognized organisms

This function returns the enriched biological processes in a set of genes, as compared to a background set of genes. The GO terms are retreived by the clusterProfiler package and the bioconductor organism databases (for example org.At.tair.db). The frequencies comparisons between GO terms in the genes of interest and the background are performed using Fischer tests, with 0.05 adjusted pvalues threshold. Those significantly enriched GO terms are then simplified with a semantic similarity threshold that can be chosen. It returns a dataframe containing, for each significant GO, their gene count, gene symbols, adjusted pvalue, description, etc.

enrich_go(
  genes,
  background,
  org = org.At.tair.db::org.At.tair.db,
  sim_cutoff = 0.85,
  GO_type = "BP",
  fdr = 0.05,
  pval = 0.05
)

Arguments

genes

gene set of interest. MUST be entrez IDs, that can be obtained from DIANE's functions convert_from_X(gene_IDs) depending on your organism. If your gene IDs are splicing aware, remove this information before with get_locus(gene_IDs).

background

gene set considered as the "universe" against which perfom the tests. MUST be entrez IDS that can be obtained from DIANE's functions convert_from_X(gene_IDs) depending on your organism. If your gene IDs are splicing aware, remove this information before with get_locus(gene_IDs).

org

organism, in the format of bioconductor organisms databases (e.g org.xx.xx.db). Possible values are, for DIANE's recognized organisms : org.Mm.eg.db, org.Hs.eg.db, org.Dm.eg.db, org.Ce.eg.db, org.EcK12.eg.db, org.At.tair.db

sim_cutoff

similarity cutoff to use for pooling similar GO terms

GO_type

character between BP (Biological Process), CC(Cellular Component) or MF (Molecular Function), depending on the GO subtype to use in the analysis

fdr

adjusted pvalues threshold, default id 0.05

pval

pvalues threshold, default id 0.05

Value

data.frame with gene significant gene ontologies, and their characteristics as columns

Examples

data("abiotic_stresses")
genes <- abiotic_stresses$heat_DEGs

genes <- get_locus(genes)
background <- get_locus(rownames(abiotic_stresses$normalized_counts))

genes <- convert_from_agi(genes)
background <- convert_from_agi(background)

go <- enrich_go(genes, background)
head(go)
#>                    ID                         Description GeneRatio   BgRatio
#> GO:0009408 GO:0009408                    response to heat    58/594 282/18247
#> GO:0006457 GO:0006457                     protein folding    39/594 192/18247
#> GO:0001666 GO:0001666                 response to hypoxia    30/594 300/18247
#> GO:0036293 GO:0036293 response to decreased oxygen levels    30/594 307/18247
#> GO:0070482 GO:0070482           response to oxygen levels    30/594 308/18247
#> GO:0071456 GO:0071456        cellular response to hypoxia    25/594 227/18247
#>                  pvalue     p.adjust       qvalue
#> GO:0009408 4.470859e-30 5.351618e-27 5.289732e-27
#> GO:0006457 2.417970e-20 1.447155e-17 1.430420e-17
#> GO:0001666 5.364220e-08 5.837247e-06 5.769745e-06
#> GO:0036293 8.967316e-08 7.694758e-06 7.605776e-06
#> GO:0070482 9.636581e-08 7.694758e-06 7.605776e-06
#> GO:0071456 1.096066e-07 7.694758e-06 7.605776e-06
#>                                                                                                                                                                                                                                                                                                                                                                                                                                  geneID
#> GO:0009408 Hop3/ATFKBP65/HSP18.5/HSP17.6B/ATHSFA2/ATHS83/ATHSP101/ATMBF1C/DjC53/ATHSP17.4/AT-HSP17.6A/HSP70T-2/ATBAG6/HSP17.6A/HSP17.8/HKL/NA/ATFKBP62/DNAJ/CLPB-M/ATERDJ3A/ATHSP70/HSP17.6II/HSP17.6C/CPN10/AtHsp90.5/HSP60/HSC70-5/HSP17.4B/HSP23.5/AT-P5C1/HSP15.7/SES1/Fes1A/AtHsp90.2/HSA32/BIP1/BIP/AR192/DREB2/AtGolS1/ATHSP22.0/ATEGY3/NA/PUB48/HSP21/NA/HSP18.2/Hsp70b/AT-HSFA3/PPI2/HTT1/DREB2/AtPARK13/AtERF53/NA/ERF54/BIP3
#> GO:0006457                                                                                                                           Hop3/ATFKBP65/HSP18.5/HSP17.6B/NA/ATHS83/ATHSP17.4/AT-HSP17.6A/HSP70T-2/ATBAG6/HSP17.6A/HSP17.8/ATFKBP62/DNAJ/CLPB-M/NA/ATHSP70/HSP17.6II/HSP17.6C/CPN10/AtHsp90.5/HSP60/HSC70-5/AtHsp90-6/HSP17.4B/HSP15.7/SES1/AtHsp90.2/ATCNX1/BIP1/BIP/AR192/Cpn60beta2/CYP40/ATHSP22.0/NA/HSP18.2/Hsp70b/BIP3
#> GO:0001666                                                                                                                                                                                                                     ATHSFA2/ATHSP101/ATMBF1C/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HKL/HSP17.6II/AT-HSFB2B/HSP17.4B/HSP15.7/ETR2/DREB2/NA/MPSR1/ATHSP22.0/HUP26/NA/AtTZF5/STMP5/ATL92/NA/NIMIN-2/NA/TN9/CYP81F2/NA/HUP32/NA
#> GO:0036293                                                                                                                                                                                                                     ATHSFA2/ATHSP101/ATMBF1C/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HKL/HSP17.6II/AT-HSFB2B/HSP17.4B/HSP15.7/ETR2/DREB2/NA/MPSR1/ATHSP22.0/HUP26/NA/AtTZF5/STMP5/ATL92/NA/NIMIN-2/NA/TN9/CYP81F2/NA/HUP32/NA
#> GO:0070482                                                                                                                                                                                                                     ATHSFA2/ATHSP101/ATMBF1C/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HKL/HSP17.6II/AT-HSFB2B/HSP17.4B/HSP15.7/ETR2/DREB2/NA/MPSR1/ATHSP22.0/HUP26/NA/AtTZF5/STMP5/ATL92/NA/NIMIN-2/NA/TN9/CYP81F2/NA/HUP32/NA
#> GO:0071456                                                                                                                                                                                                                                                ATHSFA2/ATHSP101/ATMBF1C/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HKL/HSP17.6II/AT-HSFB2B/HSP17.4B/HSP15.7/ETR2/DREB2/NA/MPSR1/ATHSP22.0/HUP26/STMP5/NA/TN9/CYP81F2/NA/HUP32/NA
#>            Count
#> GO:0009408    58
#> GO:0006457    39
#> GO:0001666    30
#> GO:0036293    30
#> GO:0070482    30
#> GO:0071456    25