Estimates Differentially Expression Genes

Given a glmFit model, performs log ratio tests using edgeR function glmLRT and returns the genes that are differentially expressed between the reference and perturbation conditions. log ratios are expressed as log(perturbation/reference)

estimateDEGs(fit, reference, perturbation, p.value = 1, lfc = 0)

Arguments

fit
edgeR glmFit

reference
condition being considered as the reference for differential analysis. It should corresponds to a condition name, e.g. the string before the underscore and replicate number in your sample names.

perturbation
condition we compared to the reference for differential analysis. It should corresponds to a condition name, e.g. the string before the underscore and replicate number in your sample names.

p.value
numeric cutoff value for adjusted p-values. Only tags with adjusted p-values equal or lower than specified are returned

lfc
minimal absolute log fold change required for a gene to be considered as differentially expressed.

Value

topTags object, which table element contains DEGs dataframe.

Examples

data("abiotic_stresses")
tcc_object <- DIANE::normalize(abiotic_stresses$raw_counts, 
abiotic_stresses$conditions, iteration = FALSE)
threshold = 10*length(abiotic_stresses$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
fit <- DIANE::estimateDispersion(tcc = tcc_object, 
conditions = abiotic_stresses$conditions)

#> Warning: norm factors don't multiply to 1

topTags <- DIANE::estimateDEGs(fit, reference = "C", perturbation = "H", p.value = 0.01)
DEGs <- topTags$table
head(DEGs)

#>                   genes    logFC   logCPM       LR        PValue           FDR
#> AT4G12400.2 AT4G12400.2 7.756985 7.564179 2272.130  0.000000e+00  0.000000e+00
#> AT5G48570.1 AT5G48570.1 7.067740 8.853736 1970.131  0.000000e+00  0.000000e+00
#> AT5G12110.1 AT5G12110.1 5.939006 7.952267 2202.123  0.000000e+00  0.000000e+00
#> AT5G64510.1 AT5G64510.1 5.879518 6.067228 2228.044  0.000000e+00  0.000000e+00
#> AT2G19310.1 AT2G19310.1 4.212070 7.686615 1389.691 3.651876e-304 1.406410e-300
#> AT2G29500.1 AT2G29500.1 7.731411 9.482692 1260.703 3.919280e-276 1.257828e-272