Get genes information

Gives gene information (common name and description) for a specific organism

get_gene_information(ids, organism)

Arguments

ids
vector of genes, AGI for Arabidopsis and ensembl for Human

organism
value must be between "Arabidopsis thaliana", "Homo sapiens", "Mus musculus", "Caenorhabditis elegans", "Escherichia coli", "Drosophilia melanogaster"

Value

Dataframe with input genes as rownames, with their label and description as columns

Examples

genes <-  c("AT2G05940", "AT4G16480", "AT4G04570", "AT2G30130", "AT1G56300")
get_gene_information(genes, organism = "Arabidopsis thaliana")

#>                                  label
#> AT2G05940                         RIPK
#> AT4G16480                         INT4
#> AT4G04570 RECEPTOR-LIKE PROTEIN KINASE
#> AT2G30130                         ASL5
#> AT1G56300                        DJC53
#>                                                                                                                                                                                                                        description
#> AT2G05940                                                                   Encodes a receptor-like cytoplasmic kinase that phosphorylates the host target RIN4, leading to the activation of a plant innate immune receptor RPM1.
#> AT4G16480                  Encodes a high affinity H+:<i>myo</i>-inositol symporter. The only other compound shown to be transported was pinitol, a methylated derivative of <i>myo</i>-inositol. The mRNA is cell-to-cell mobile.
#> AT4G04570                                                                                                                                   Encodes a cysteine-rich receptor-like protein kinase. The mRNA is cell-to-cell mobile.
#> AT2G30130 Overexpression/activation tagged allele has epinastic leaves, reduced apical dominance and is sterile. Gene is similar to asymmetric leaves (AS)/lateral organ boundary (LOB) genes  which repress KNOX gene expression.
#> AT1G56300                                                                                                                                                                                Chaperone DnaJ-domain superfamily protein


get_gene_information(c("WBGene00000013", "WBGene00000035"), 
organism = "Caenorhabditis elegans")

#>                label                            description
#> WBGene00000013 abf-2 Antibacterial factor-related peptide 2
#> WBGene00000035 ace-1                 Acetylcholinesterase 1