List of regulator genes for a number of model organisms

Those particular genes are known to be transcriptionnal actors, and are needed to infer gene regulatory networks in DIANE.

regulators_per_organism

Format

A named list with the following elements:

Arabidopsis thaliana

Character vector of the known transcriptional regulators in Arabidopsis thaliana

Homo sapiens

Character vector of the known transcription factors in Human

Mus musculus

Character vector of the known transcription factors in Mouse

Drosophilia melanogaster

Character vector of the known transcription factors in Drosophilia melanogaster

Caenorhabditis elegans

Character vector of the known transcription factors in C elegans

Escherichia coli

Character vector of the known transcription factors and sigma elements in E coli K12

Source

For Arabidopsis thaliana

http://plntfdb.bio.uni-potsdam.de/v3.0/index.php?sp_id=ATH

For Human

http://humantfs.ccbr.utoronto.ca/download.php

Mus musculus

http://bioinfo.life.hust.edu.cn/AnimalTFDB/#!/download

Drosophilia melanogaster

https://flybase.org/reports/FBgg0000745.html

C elegans

http://www.wormbook.org/chapters/www_transcriptionalregulation.2/transregulate.html#sec5-1

E coli

http://regulondb.ccg.unam.mx/menu/download/datasets/index.jsp

Examples

{
 print(head(regulators_per_organism[["Arabidopsis thaliana"]]))
 print(head(regulators_per_organism[["Homo sapiens"]]))
 print(head(regulators_per_organism[["Mus musculus"]]))
 print(head(regulators_per_organism[["Drosophilia melanogaster"]]))
 print(head(regulators_per_organism[["Caenorhabditis elegans"]]))
 print(head(regulators_per_organism[["Escherichia coli"]]))
}
#> [1] "AT1G04880" "AT1G20910" "AT1G55650" "AT1G76110" "AT1G76510" "AT2G17410"
#> [1] "ENSG00000004848" "ENSG00000005073" "ENSG00000005102" "ENSG00000005513"
#> [5] "ENSG00000005801" "ENSG00000005889"
#> [1] "ENSMUSG00000037029" "ENSMUSG00000032035" "ENSMUSG00000021485"
#> [4] "ENSMUSG00000097200" "ENSMUSG00000098022" "ENSMUSG00000027985"
#> [1] "FBgn0002522" "FBgn0051481" "FBgn0004054" "FBgn0004053" "FBgn0000166"
#> [6] "FBgn0000439"
#> [1] "WBGene00019424" "WBGene00009202" "WBGene00009203" "WBGene00013383"
#> [5] "WBGene00000476" "WBGene00002717"
#> [1] "accB" "acrR" "ada"  "adiY" "agaR" "aidB"