run_coseq

run_coseq

run_coseq(conds, genes, data, K = 6:12)

Arguments

conds
Condition names to be used for clustering

genes
Genes used as an input for the clustering

data
normalized counts (MUST be normalized!)

K
number of clusters range

Value

named list containing the coseq run result as "model", and the cluster membership for each gene as "membership"

Examples

data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)

#> TCC::INFO: Calculating normalization factors using tmm ...

#> TCC::INFO: Done.

threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

#> Warning: norm factors don't multiply to 1

topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes
clustering <- DIANE::run_coseq(conds = unique(demo_data_At$conditions),
data = normalized_counts, genes = genes, K = 6:9)

#> ****************************************
#> coseq analysis: Poisson approach & none transformation
#> K = 6 to 9 
#> Use set.seed() prior to running coseq for reproducible results.
#> ****************************************
#> Running K = 6 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 9.83259654335987"
#> [1] "Log-like diff: 14.1699216327095"
#> [1] "Log-like diff: 10.0395316306507"
#> [1] "Log-like diff: 10.336858994218"
#> [1] "Log-like diff: 14.052428610394"

#> Running K = 7 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 144.636208534122"
#> [1] "Log-like diff: 84.1884845728268"
#> [1] "Log-like diff: 183.80722199113"
#> [1] "Log-like diff: 106.821522550999"
#> [1] "Log-like diff: 56.1969685492668"

#> Running K = 8 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 16.2420432975132"
#> [1] "Log-like diff: 8.54152519123873"
#> [1] "Log-like diff: 4.23446140731638"
#> [1] "Log-like diff: 1.87643946402148"
#> [1] "Log-like diff: 0.637900202910378"

#> Running K = 9 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 0.900924955526349"
#> [1] "Log-like diff: 0.296311784087115"
#> [1] "Log-like diff: 0.0397224670583611"
#> [1] "Log-like diff: 0.00348793552894833"
#> [1] "Log-like diff: 0.00686927938164317"

#> Note: slope heuristics for model selection only applied if > 10 models are fit.