sample_correlation

Correlation matrix between samples, computed on log2(counts + 1). Shared by the static and interactive correlation heatmaps.

Expects counts, raw or normalised, never already transformed : the log is applied here. Note that it only affects pearson : spearman and kendall work on ranks, which a monotonic transform leaves untouched.

sample_correlation(
  data,
  conds = unique(stringr::str_split_fixed(colnames(data), "_", 2)[, 1]),
  correlation_method = "pearson"
)

Arguments

data

a matrix of counts, samples as columns. At least two samples and three genes, no negative value.

conds

conditions to keep. Default : all of them. Conditions that are not in the data are ignored, with a warning.

correlation_method

correlation method for the cor function. Can be pearson, kendall or spearman.

Value

a square, unrounded correlation matrix

Examples

data("abiotic_stresses")
sample_correlation(abiotic_stresses$raw_counts)
#>             C_1       C_2       C_3       S_1       S_2       S_3       M_1
#> C_1   1.0000000 0.9931175 0.9916752 0.9890897 0.9903233 0.9907028 0.9878956
#> C_2   0.9931175 1.0000000 0.9927650 0.9898597 0.9901993 0.9902852 0.9878864
#> C_3   0.9916752 0.9927650 1.0000000 0.9878910 0.9892149 0.9890720 0.9889024
#> S_1   0.9890897 0.9898597 0.9878910 1.0000000 0.9919296 0.9918539 0.9875690
#> S_2   0.9903233 0.9901993 0.9892149 0.9919296 1.0000000 0.9933528 0.9880249
#> S_3   0.9907028 0.9902852 0.9890720 0.9918539 0.9933528 1.0000000 0.9879237
#> M_1   0.9878956 0.9878864 0.9889024 0.9875690 0.9880249 0.9879237 1.0000000
#> M_2   0.9875260 0.9866873 0.9863162 0.9869668 0.9864467 0.9868928 0.9922657
#> M_3   0.9861962 0.9856710 0.9839829 0.9875783 0.9865642 0.9869201 0.9910177
#> H_1   0.9737834 0.9736532 0.9729691 0.9759361 0.9764902 0.9761437 0.9727372
#> H_2   0.9695380 0.9681951 0.9665135 0.9722727 0.9728314 0.9724626 0.9674471
#> H_3   0.9706344 0.9710796 0.9704475 0.9736008 0.9743343 0.9734692 0.9703551
#> SM_1  0.9843017 0.9842188 0.9811934 0.9876596 0.9859127 0.9863650 0.9872996
#> SM_2  0.9861155 0.9859692 0.9836581 0.9888571 0.9862113 0.9867862 0.9888551
#> SM_3  0.9873100 0.9871911 0.9846289 0.9892100 0.9873559 0.9878962 0.9889583
#> SH_1  0.9679045 0.9685181 0.9681179 0.9723051 0.9728500 0.9718584 0.9692918
#> SH_2  0.9678784 0.9679792 0.9675823 0.9730142 0.9732599 0.9724106 0.9694662
#> SH_3  0.9719399 0.9711994 0.9704297 0.9759935 0.9767805 0.9763833 0.9721409
#> MH_1  0.9710037 0.9712231 0.9697292 0.9751137 0.9746128 0.9741479 0.9735782
#> MH_2  0.9627121 0.9638595 0.9617828 0.9708786 0.9675664 0.9678174 0.9664937
#> MH_3  0.9709433 0.9712343 0.9709878 0.9747586 0.9744083 0.9740980 0.9755164
#> SMH_1 0.9694808 0.9694933 0.9661352 0.9758644 0.9738231 0.9739594 0.9707540
#> SMH_2 0.9671860 0.9668152 0.9641142 0.9737568 0.9723175 0.9722688 0.9700623
#> SMH_3 0.9657683 0.9659998 0.9637779 0.9732839 0.9722674 0.9719984 0.9681376
#>             M_2       M_3       H_1       H_2       H_3      SM_1      SM_2
#> C_1   0.9875260 0.9861962 0.9737834 0.9695380 0.9706344 0.9843017 0.9861155
#> C_2   0.9866873 0.9856710 0.9736532 0.9681951 0.9710796 0.9842188 0.9859692
#> C_3   0.9863162 0.9839829 0.9729691 0.9665135 0.9704475 0.9811934 0.9836581
#> S_1   0.9869668 0.9875783 0.9759361 0.9722727 0.9736008 0.9876596 0.9888571
#> S_2   0.9864467 0.9865642 0.9764902 0.9728314 0.9743343 0.9859127 0.9862113
#> S_3   0.9868928 0.9869201 0.9761437 0.9724626 0.9734692 0.9863650 0.9867862
#> M_1   0.9922657 0.9910177 0.9727372 0.9674471 0.9703551 0.9872996 0.9888551
#> M_2   1.0000000 0.9928373 0.9726897 0.9690199 0.9698683 0.9889555 0.9916310
#> M_3   0.9928373 1.0000000 0.9717586 0.9694552 0.9687388 0.9925789 0.9924123
#> H_1   0.9726897 0.9717586 1.0000000 0.9909936 0.9931471 0.9694286 0.9710259
#> H_2   0.9690199 0.9694552 0.9909936 1.0000000 0.9915685 0.9672781 0.9681334
#> H_3   0.9698683 0.9687388 0.9931471 0.9915685 1.0000000 0.9665798 0.9685258
#> SM_1  0.9889555 0.9925789 0.9694286 0.9672781 0.9665798 1.0000000 0.9931530
#> SM_2  0.9916310 0.9924123 0.9710259 0.9681334 0.9685258 0.9931530 1.0000000
#> SM_3  0.9909133 0.9927802 0.9714700 0.9687438 0.9685091 0.9934362 0.9935888
#> SH_1  0.9678063 0.9665654 0.9912423 0.9872996 0.9914684 0.9651982 0.9665410
#> SH_2  0.9682132 0.9677739 0.9911978 0.9890182 0.9918001 0.9665029 0.9674189
#> SH_3  0.9719920 0.9718678 0.9922198 0.9909579 0.9918848 0.9703272 0.9713931
#> MH_1  0.9739094 0.9741422 0.9916385 0.9879075 0.9906015 0.9729101 0.9734338
#> MH_2  0.9690924 0.9690931 0.9873524 0.9854222 0.9866169 0.9683148 0.9703812
#> MH_3  0.9753558 0.9739627 0.9911181 0.9868698 0.9906765 0.9711886 0.9732760
#> SMH_1 0.9727622 0.9754443 0.9875694 0.9865253 0.9863651 0.9770008 0.9757471
#> SMH_2 0.9721367 0.9742950 0.9871854 0.9873840 0.9863986 0.9755445 0.9744872
#> SMH_3 0.9690436 0.9712405 0.9886567 0.9877667 0.9881685 0.9714897 0.9706417
#>            SM_3      SH_1      SH_2      SH_3      MH_1      MH_2      MH_3
#> C_1   0.9873100 0.9679045 0.9678784 0.9719399 0.9710037 0.9627121 0.9709433
#> C_2   0.9871911 0.9685181 0.9679792 0.9711994 0.9712231 0.9638595 0.9712343
#> C_3   0.9846289 0.9681179 0.9675823 0.9704297 0.9697292 0.9617828 0.9709878
#> S_1   0.9892100 0.9723051 0.9730142 0.9759935 0.9751137 0.9708786 0.9747586
#> S_2   0.9873559 0.9728500 0.9732599 0.9767805 0.9746128 0.9675664 0.9744083
#> S_3   0.9878962 0.9718584 0.9724106 0.9763833 0.9741479 0.9678174 0.9740980
#> M_1   0.9889583 0.9692918 0.9694662 0.9721409 0.9735782 0.9664937 0.9755164
#> M_2   0.9909133 0.9678063 0.9682132 0.9719920 0.9739094 0.9690924 0.9753558
#> M_3   0.9927802 0.9665654 0.9677739 0.9718678 0.9741422 0.9690931 0.9739627
#> H_1   0.9714700 0.9912423 0.9911978 0.9922198 0.9916385 0.9873524 0.9911181
#> H_2   0.9687438 0.9872996 0.9890182 0.9909579 0.9879075 0.9854222 0.9868698
#> H_3   0.9685091 0.9914684 0.9918001 0.9918848 0.9906015 0.9866169 0.9906765
#> SM_1  0.9934362 0.9651982 0.9665029 0.9703272 0.9729101 0.9683148 0.9711886
#> SM_2  0.9935888 0.9665410 0.9674189 0.9713931 0.9734338 0.9703812 0.9732760
#> SM_3  1.0000000 0.9664843 0.9675055 0.9716300 0.9733256 0.9689680 0.9725362
#> SH_1  0.9664843 1.0000000 0.9916761 0.9905163 0.9905765 0.9864187 0.9907134
#> SH_2  0.9675055 0.9916761 1.0000000 0.9922790 0.9899968 0.9862613 0.9903950
#> SH_3  0.9716300 0.9905163 0.9922790 1.0000000 0.9906208 0.9868289 0.9906946
#> MH_1  0.9733256 0.9905765 0.9899968 0.9906208 1.0000000 0.9900433 0.9929632
#> MH_2  0.9689680 0.9864187 0.9862613 0.9868289 0.9900433 1.0000000 0.9893969
#> MH_3  0.9725362 0.9907134 0.9903950 0.9906946 0.9929632 0.9893969 1.0000000
#> SMH_1 0.9762732 0.9861357 0.9868864 0.9880769 0.9908025 0.9892392 0.9883661
#> SMH_2 0.9748350 0.9855847 0.9872428 0.9886728 0.9901834 0.9898151 0.9886784
#> SMH_3 0.9712169 0.9879725 0.9893395 0.9899775 0.9905666 0.9894695 0.9888613
#>           SMH_1     SMH_2     SMH_3
#> C_1   0.9694808 0.9671860 0.9657683
#> C_2   0.9694933 0.9668152 0.9659998
#> C_3   0.9661352 0.9641142 0.9637779
#> S_1   0.9758644 0.9737568 0.9732839
#> S_2   0.9738231 0.9723175 0.9722674
#> S_3   0.9739594 0.9722688 0.9719984
#> M_1   0.9707540 0.9700623 0.9681376
#> M_2   0.9727622 0.9721367 0.9690436
#> M_3   0.9754443 0.9742950 0.9712405
#> H_1   0.9875694 0.9871854 0.9886567
#> H_2   0.9865253 0.9873840 0.9877667
#> H_3   0.9863651 0.9863986 0.9881685
#> SM_1  0.9770008 0.9755445 0.9714897
#> SM_2  0.9757471 0.9744872 0.9706417
#> SM_3  0.9762732 0.9748350 0.9712169
#> SH_1  0.9861357 0.9855847 0.9879725
#> SH_2  0.9868864 0.9872428 0.9893395
#> SH_3  0.9880769 0.9886728 0.9899775
#> MH_1  0.9908025 0.9901834 0.9905666
#> MH_2  0.9892392 0.9898151 0.9894695
#> MH_3  0.9883661 0.9886784 0.9888613
#> SMH_1 1.0000000 0.9930953 0.9917508
#> SMH_2 0.9930953 1.0000000 0.9918143
#> SMH_3 0.9917508 0.9918143 1.0000000