Combined abiotic stresses

This dataset contians the transcriptome of Arabidopsis thaliana plants exposed to global warming induced conditions. It was generated in the article "Molecular plant responses to combined abiotic stresses put a spotlight on unknown and abundant genes", by Sewelam et Al. in Journal of experimental Botany, 2020. The experimental perturbations studied are high tempreature, hight salinity and osmotic changes in the soil. Each factors has two levels, one of them considered as the reference, and the other one as the stress level.

abiotic_stresses

Format

A named list with the following elements:

raw_counts

Dataframe of positive values. Raw transcript aboundances as obtained after mapping and quantifying RNASeq reads. Rows are transcripts, and columns are experimental triplicate conditions

design

Dataframe. Describes for each condition, the level of each factor.

conditions

Character vector. Gives the condition corresponding to each column of the raw_counts element

Source

Data download :

GEO accession
Publication :

Molecular plant responses to combined abiotic stresses put a spotlight on unknown and abundant genes
Examples

{
 head(abiotic_stresses$raw_counts)
 abiotic_stresses$design
 abiotic_stresses$conditions
}

#>  [1] "C"   "C"   "C"   "S"   "S"   "S"   "M"   "M"   "M"   "H"   "H"   "H"  
#> [13] "SM"  "SM"  "SM"  "SH"  "SH"  "SH"  "MH"  "MH"  "MH"  "SMH" "SMH" "SMH"