Targets and regulators of a gene in a network

Targets and regulators of a gene in a network

describe_node(graph, node)

Arguments

graph
igraph object

node
gene to describe

Value

list with targets and regulators attributes

Examples

data("abiotic_stresses")
data("regulators_per_organism")
mat <- abiotic_stresses$heat_DEGs_regulatory_links
network <- DIANE::network_thresholding(mat, n_edges = 600)
data <- network_data(network, regulators_per_organism[["Arabidopsis thaliana"]])
nodes <- data$nodes[order(-data$nodes$degree),]
gene_of_interest <- nodes$id[1]
DIANE::describe_node(network, node =gene_of_interest)

#> $targets
#>  [1] "AT3G12910" "AT5G24110" "AT3G23230" "AT5G56960" "AT1G44830" "AT2G43590"
#>  [7] "AT1G75040" "AT1G60750" "AT4G11470" "AT3G57240" "AT1G35230" "AT4G18253"
#> [13] "AT5G57220" "AT4G39610" "AT1G21120" "AT1G16090" "AT1G05675" "AT1G56240"
#> [19] "AT3G53250" "AT5G25910" "AT4G18250" "AT1G15125" "AT4G19975" "AT2G37980"
#> [25] "AT4G19430" "AT4G25070" "AT3G23550" "AT3G20395" "AT5G54710" "AT1G54040"
#> [31] "AT5G60310"
#> 
#> $regulators
#> [1] "AT3G12910" "AT5G24110" "AT3G23230" "AT5G56960" "AT1G01060"
#>