Draw network with removed edges in red

Draw network with removed edges in red

draw_discarded_edges(links, net_data)

Arguments

links

dataframe of edges containing their adjusted pvalues, as returned by the test_edges function

net_data

network data of the thresholded network, given by DIANE::network_data()

Examples

data(abiotic_stresses)
links <- abiotic_stresses$heat_edge_tests$links
net <- network_from_tests(links, fdr = 0.01)
#> 0 edges kept in final network
net_data <- network_data(net, 
gene_info = gene_annotations$`Arabidopsis thaliana`, 
regulators = regulators_per_organism$`Arabidopsis thaliana`)
draw_discarded_edges(links, net_data)
#> 475 edges kept in final network