Computes enrich Gene Ontology terms in a set of genes

This function returns the enriched biological processes in a set of genes, as compared to a background set of genes. The GO terms are retreived by the clusterProfiler package and the bioconductor organism databases (for example org.At.tair.db). The frequencies comparisons between GO terms in the genes of interest and the background are performed using Fischer tests, with 0.05 adjusted pvalues threshold. Those significantly enriched GO terms are then simplified with a semantic similarity threshold that can be chosen. It returns a dataframe containing, for each significant GO, their gene count, gene symbols, adjusted pvalue, description, etc.

enrich_go(
  genes,
  background,
  org = org.At.tair.db::org.At.tair.db,
  sim_cutoff = 0.85
)

Arguments

genes
gene set of interest. MUST be entrez IDs.

background
gene set considered as the "universe" against which perfom the tests. MUST be entrez IDS.

org
organism, in the format of bioconductor organisms databases (e.g org.xx.xx.db)

sim_cutoff
similarity cutoff to use for pooling similar GO terms

Value

data.frame

Examples

data("abiotic_stresses")
genes <- abiotic_stresses$heat_DEGs

genes <- get_locus(genes)
background <- get_locus(rownames(abiotic_stresses$normalized_counts))

genes <- convert_from-agi(genes)

#> Error in eval(expr, envir, enclos): object 'convert_from' not found

background <- convert_from_agi(background)

go <- enrich_go(genes, background)

#> 

#> --> No gene can be mapped....

#> --> Expected input gene ID: 821972,822493,832453,831436,841165,841684

#> --> return NULL...

head(go)

#> NULL