Computes enrich Gene Ontology terms in a set of genes

This function returns the enriched biological processes in a set of genes, as compared to a background set of genes. The GO terms are retreived by the clusterProfiler package and the bioconductor organism databases (for example org.At.tair.db). The frequencies comparisons between GO terms in the genes of interest and the background are performed using Fischer tests, with 0.05 adjusted pvalues threshold. Those significantly enriched GO terms are then simplified with a semantic similarity threshold that can be chosen. It returns a dataframe containing, for each significant GO, their gene count, gene symbols, adjusted pvalue, description, etc.

enrich_go(
  genes,
  background,
  org = org.At.tair.db::org.At.tair.db,
  sim_cutoff = 0.85
)

Arguments

genes
gene set of interest. MUST be entrez IDs.

background
gene set considered as the "universe" against which perfom the tests. MUST be entrez IDS.

org
organism, in the format of bioconductor organisms databases (e.g org.xx.xx.db)

sim_cutoff
similarity cutoff to use for pooling similar GO terms

Value

data.frame

Examples

data("abiotic_stresses")
genes <- abiotic_stresses$heat_DEGs

genes <- get_locus(genes)
background <- get_locus(rownames(abiotic_stresses$normalized_counts))

genes <- convert_from_agi(genes)
background <- convert_from_agi(background)

go <- enrich_go(genes, background)

#> 

head(go)

#>                    ID                  Description GeneRatio   BgRatio
#> GO:0009408 GO:0009408             response to heat    52/498 211/15722
#> GO:0006457 GO:0006457              protein folding    37/498 170/15722
#> GO:0042493 GO:0042493             response to drug    36/498 423/15722
#> GO:0006970 GO:0006970   response to osmotic stress    36/498 490/15722
#> GO:0006979 GO:0006979 response to oxidative stress    30/498 405/15722
#> GO:0071456 GO:0071456 cellular response to hypoxia    25/498 223/15722
#>                  pvalue     p.adjust       qvalue
#> GO:0009408 7.669228e-32 9.548189e-29 9.404896e-29
#> GO:0006457 6.570526e-21 4.090152e-18 4.028770e-18
#> GO:0042493 7.451098e-08 6.184411e-06 6.091599e-06
#> GO:0006970 2.555805e-06 1.272791e-04 1.253690e-04
#> GO:0006979 1.512732e-05 6.975376e-04 6.870693e-04
#> GO:0071456 4.362043e-08 4.525620e-06 4.457702e-06
#>                                                                                                                                                                                                                                                                                                                                                                     geneID
#> GO:0009408 Hop3/ATFKBP65/NA/NA/ATHSFA2/ATHS83/ATHSP101/ATMBF1C/NA/ATHSP17.4/AT-HSP17.6A/HSP70T-2/ATBAG6/NA/NA/HLP1/ATFKBP62/CLPB-M/ATERDJ3A/ATHSP70/HSP17.6II/NA/CPN10/AtHsp90.5/AT-HSFB2B/HSP60/HSC70-5/NA/AT-P5C1/NA/SES1/Fes1A/AtHsp90.2/HSA32/BIP1/BIP/AR192/DREB2/AtGolS1/ATHSP22.0/PUB48/HSP21/HSP18.2/Hsp70b/AT-HSFA3/AT-HSFB2A/HTT1/DREB2/AtPARK13/AtERF53/NA/BIP3
#> GO:0006457                                                                                                             ATFKBP65/NA/NA/NA/ATHS83/ATHSP17.4/AT-HSP17.6A/HSP70T-2/ATBAG6/NA/NA/ATFKBP62/DNAJ/CLPB-M/NA/ATHSP70/HSP17.6II/NA/CPN10/AtHsp90.5/HSP60/HSC70-5/AtHsp90-6/NA/NA/SES1/AtHsp90.2/ATCNX1/BIP1/BIP/AR192/Cpn60beta2/CYP40/ATHSP22.0/HSP18.2/Hsp70b/BIP3
#> GO:0042493                                                                                                                 NA/NA/ATHSFA2/ATHSP17.4/AT-HSP17.6A/NA/NA/HSP17.6II/NA/AT-HSFB2B/NA/NA/DREB2/ATHSP22.0/ATWRKY54/NA/PCC1/HSP18.2/ATUPS1/AT-HSFA3/AT-HSFB2A/ANAC092/ATWRKY38/LURP1/NA/ARR7/AtERF98/ATWRKY30/AtRLP52/ATFER1/NA/AtERF014/ZAT11/AtSWEET15/AtCAPE9/NA
#> GO:0006970                                                                                                           NA/NA/ATGRP7/ATHSP17.4/AT-HSP17.6A/NA/NA/ATFKBP62/HSP17.6II/NA/AtHsp90.5/NA/PIP2;3/AT-P5C1/NA/Fes1A/AtHsp90.2/ABX45/AtGolS1/ATHSP22.0/ATWRKY54/HSP18.2/RAP2.4/NA/ANAC092/AtLtpI-11/CCX2/5PTASE11/AtERF53/AtLEA3/FAR5/AtRLP30/NA/ERD7/AtSWEET15/ATTSPO
#> GO:0006979                                                                                                                                                                                 NA/NA/ATHSFA2/ATHSP17.4/AT-HSP17.6A/NA/NA/HSP17.6II/NA/NA/NA/NA/DREB2/AtGolS1/ATHSP22.0/NA/APX1B/ATWRKY54/AtGolS3/HSP18.2/NA/NA/ANAC092/PRX40/ATFER3/PRX9/ATWRKY30/ATFER1/NA/NA
#> GO:0071456                                                                                                                                                                                                                  ATHSFA2/ATHSP101/ATMBF1C/ATHSP17.4/AT-HSP17.6A/NA/NA/HLP1/HSP17.6II/AT-HSFB2B/NA/NA/ETR2/DREB2/NA/MPSR1/ATHSP22.0/NA/NA/NA/NA/CYP81F2/NA/NA/NA
#>            Count
#> GO:0009408    52
#> GO:0006457    37
#> GO:0042493    36
#> GO:0006970    36
#> GO:0006979    30
#> GO:0071456    25