Computes enrich Gene Ontology terms in a set of genes

This function returns the enriched biological processes in a set of genes, as compared to a background set of genes. The GO terms are retreived by the clusterProfiler package and the bioconductor organism databases (for example org.At.tair.db). The frequencies comparisons between GO terms in the genes of interest and the background are performed using Fischer tests, with 0.05 adjusted pvalues threshold. Those significantly enriched GO terms are then simplified with a semantic similarity threshold that can be chosen. It returns a dataframe containing, for each significant GO, their gene count, gene symbols, adjusted pvalue, description, etc.

enrich_go(
  genes,
  background,
  org = org.At.tair.db::org.At.tair.db,
  sim_cutoff = 0.85
)

Arguments

genes
gene set of interest. MUST be entrez IDs.

background
gene set considered as the "universe" against which perfom the tests. MUST be entrez IDS.

org
organism, in the format of bioconductor organisms databases (e.g org.xx.xx.db)

sim_cutoff
similarity cutoff to use for pooling similar GO terms

Value

data.frame

Examples

data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)

#> TCC::INFO: Calculating normalization factors using tmm ...

#> TCC::INFO: Done.

threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

#> Warning: norm factors don't multiply to 1

topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
# interest and background gene sets as entrez ids
genes <- convert_from_agi(topTags$table$genes)
background <- convert_from_agi(rownames(normalized_counts))
go <- enrich_go(genes, background)
head(go)

#>                    ID                   Description GeneRatio   BgRatio
#> GO:0009617 GO:0009617         response to bacterium  126/2069 477/16595
#> GO:0042493 GO:0042493              response to drug  119/2069 433/16595
#> GO:0042742 GO:0042742 defense response to bacterium  115/2069 413/16595
#> GO:0002376 GO:0002376         immune system process  103/2069 393/16595
#> GO:0006955 GO:0006955               immune response   95/2069 357/16595
#> GO:0045087 GO:0045087        innate immune response   93/2069 343/16595
#>                  pvalue     p.adjust       qvalue
#> GO:0009617 4.536259e-17 1.531743e-14 1.404648e-14
#> GO:0042493 1.327968e-17 7.854222e-15 7.202526e-15
#> GO:0042742 1.627326e-17 7.854222e-15 7.202526e-15
#> GO:0002376 6.189334e-14 1.044966e-11 9.582609e-12
#> GO:0006955 2.232723e-13 3.231069e-11 2.962974e-11
#> GO:0045087 1.208270e-13 1.883043e-11 1.726799e-11
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         geneID
#> GO:0009617 FMO1/ALD1/SARD1/ATICS1/ATIRT1/ATGST11/HR4/NA/DLO1/LecRK-I.7/ATWRKY54/CRK5/ATFRO2/NA/CBP60G/ATWRKY70/ATWRKY51/FRK1/AZI1/ATL31/LecRK-VI.3/NHL3/CYP71A12/BAL/ATGSTF2/AT-RLK3/SIB1/AT-POX/LecRK-III.2/SOBER1/NA/Nog1-2/AtHsp70-2/ATHSPRO2/LecRK-I.8/AtNUDT7/GLIP3/ZAR1/ATGST1/ATMYB72/AtDMR6/RIPK/ATPCAP1/ATOSM34/LecRK-II.2/HWI1/LecRK-V.9/ATCHITIV/AIG1/CYP79B2/PCRK1/ER/ALPHA-DOX1/PR-1-LIKE/NA/AGO2/NIMIN-1/ATWRKY48/LecRK-III.1/CYP71A13/ATNPR3/LecRK-IV.1/ATWRKY53/MEE54/LecRK-V.2/TAO1/LecRK-I.2/LSU2/XLG2/ATL6/LecRK-V.3/ADR1-L1/CYP71B15/CRPK1/AtPFA-DSP4/ATEXO70B2/AtLYK4/AtLPK1/PSL4/LecRK-V.1/LecRK-IV.2/LecRK-I.1/ATWRKY62/BAH1/LecRK-II.1/NA/IOS1/BGLU26/BIK1/LecRK-I.4/ADR1-L2/BIR1/ARA8/AtNHR2B/APK2A/AtEIN3/AtTCTP1/ATHLECRK/AtCERK1/LecRK-I.3/ATMEK1/LecRK-IX.1/GLI1/AtTTM2/mMDH1/LecRK-I.6/BIR2/ATNPR4/AtJMJ17/ATHS83/ATWRKY33/ATMYB44/GLIP1/ATPAD4/NA/NHL13/RPS4/ATR4/RTP1/RGLG4/VI1/PSL5/CYP81F2/CRK34/ATBAK1/NDR1
#> GO:0042493                                                                                                                                             ALD1/STOP2/NA/HR4/DLO1/ATWRKY54/PMZ/CRK5/RMG1/ATFSD1/NA/ATWRKY70/STZ/ATL31/WAK2/ATOPR1/ASG6/ATACP5/AtIQM1/ATBT2/BOR2/NHL3/BAL/LURP1/ATGSTU7/ATCMPG1/A/N-InvG/AtZAT12/ERD6/ATHSPRO2/ATMYB15/AtNUDT7/BT1/AtDMR6/EDR4/ATWRKY29/AT-SYR1/NA/ATSYP122/anac036/ATPUMP5/CRCK1/ALPHA-DOX1/DREB26/NA/ATWRKY48/AtWAK1/ATPUB29/ATHSFA2/BAP1/PIP2;4/NA/AZF1/LecRK-IV.1/AtBRN1/SCL13/ARR7/AtRDR1/ATWRKY53/CPK31/CAT1/anac016/ATSZF1/LHT6/AGD2/ATL6/AEP3/F6'H1/NA/ATBZIP60/AAT1/AZF2/AtLYK4/ATGRP-3/AtWRKY22/ATWRKY62/BAH1/IOS1/ATNUDT6/AtWRKY15/AtRLP52/AtPUB23/AtCERK1/NA/ATMEK1/AtTTM2/NA/AtRDUF2/AtPUB22/NA/ATNPR4/ATWRKY33/ATMYB44/ATNOA1/GLIP1/ATPAD4/AT-HSFB2B/NA/AtRDUF1/REM1.2/ATL2/NA/Rem1.3/GAPC/AT-HSP17.6A/NA/ATWRKY30/ZAT18/DREB2/MYB77/NA/NA/ATERF6/NA/NA/BASS6/SAUR21/AT-HSFA4A/AtZAT6
#> GO:0042742                                                                          FMO1/ALD1/SARD1/ATICS1/ATGST11/NA/LecRK-I.7/ATWRKY54/CRK5/NA/CBP60G/ATWRKY70/ATWRKY51/FRK1/AZI1/ATL31/LecRK-VI.3/NHL3/CYP71A12/BAL/ATGSTF2/AT-RLK3/SIB1/AT-POX/LecRK-III.2/SOBER1/NA/Nog1-2/ATHSPRO2/LecRK-I.8/AtNUDT7/GLIP3/ZAR1/ATGST1/ATMYB72/AtDMR6/RIPK/ATPCAP1/ATOSM34/LecRK-II.2/HWI1/LecRK-V.9/CYP79B2/PCRK1/ER/ALPHA-DOX1/PR-1-LIKE/NA/AGO2/NIMIN-1/ATWRKY48/LecRK-III.1/CYP71A13/ATNPR3/LecRK-IV.1/ATWRKY53/MEE54/LecRK-V.2/TAO1/LecRK-I.2/LSU2/XLG2/ATL6/LecRK-V.3/ADR1-L1/AtPFA-DSP4/ATEXO70B2/AtLPK1/PSL4/LecRK-V.1/LecRK-IV.2/LecRK-I.1/ATWRKY62/BAH1/LecRK-II.1/NA/IOS1/BGLU26/BIK1/LecRK-I.4/ADR1-L2/BIR1/ARA8/AtNHR2B/APK2A/AtEIN3/AtTCTP1/ATHLECRK/AtCERK1/LecRK-I.3/ATMEK1/LecRK-IX.1/GLI1/AtTTM2/mMDH1/LecRK-I.6/BIR2/ATNPR4/AtJMJ17/ATHS83/ATWRKY33/ATMYB44/GLIP1/ATPAD4/NA/NHL13/RPS4/ATR4/RTP1/RGLG4/PSL5/CYP81F2/CRK34/ATBAK1/NDR1
#> GO:0002376                                                                                                                                                                                                                                              FMO1/ALD1/AtGH3.12/SARD1/ATICS1/ATGST11/SGT1A/ATNHL10/CBP60G/ATWRKY70/AZI1/ATL31/NA/PLA/SARD4/NA/NIMIN-2/NA/NHL3/CYP71A12/NA/BAL/AT-RLK3/SIB1/SOBER1/ATMES1/NA/CES101/ATHSPRO2/AtNUDT7/NSL1/ZAR1/RPM1/ATMYB72/ATOSM34/ATCHITIV/AT-SYR1/CYP79B2/ATSYP122/PCRK1/ALPHA-DOX1/AGO2/NIMIN-1/ATEXO70E2/CYP71A13/ATNPR3/ATMLO12/PROPEP3/PDI/AtBRN1/ATWRKY53/MEE54/TAO1/AGD2/ATL6/CYP71B15/MKS1/AtLYK4/ATMKK2/BAH1/ATCNGC2/ATSOT16/PROPEP2/BGLU26/ATPUB17/CHI/BIK1/AtNHR2B/RPP4/NA/APUM5/AtPUB23/AtCERK1/ATMEK1/AtTTM2/AtPUB22/NA/ATNPR4/AtGILP/JAS1/ATPEPR1/ATHS83/ATWRKY33/CPR5/NA/GLIP1/ATPAD4/NA/ATHCHIB/RPP13/CAD1/RPS4/UGT71C3/REM1.2/NA/ATR4/Rem1.3/PEP7/PROPEP6/CYP81F2/PH1/ARLA1B/NDR1
#> GO:0006955                                                                                                                                                                                                                                                                                                      FMO1/ALD1/AtGH3.12/SARD1/ATICS1/ATGST11/SGT1A/CBP60G/ATWRKY70/AZI1/ATL31/NA/PLA/SARD4/NA/NIMIN-2/NA/NHL3/CYP71A12/NA/BAL/AT-RLK3/SIB1/SOBER1/ATMES1/NA/CES101/ATHSPRO2/AtNUDT7/NSL1/ZAR1/RPM1/ATMYB72/ATOSM34/ATCHITIV/AT-SYR1/CYP79B2/ATSYP122/PCRK1/ALPHA-DOX1/NIMIN-1/ATEXO70E2/CYP71A13/ATNPR3/ATMLO12/PROPEP3/PDI/AtBRN1/ATWRKY53/MEE54/TAO1/AGD2/ATL6/CYP71B15/MKS1/AtLYK4/ATMKK2/BAH1/ATCNGC2/ATSOT16/PROPEP2/BGLU26/ATPUB17/CHI/BIK1/AtNHR2B/RPP4/NA/AtCERK1/ATMEK1/AtTTM2/NA/ATNPR4/AtGILP/JAS1/ATPEPR1/ATHS83/ATWRKY33/CPR5/NA/GLIP1/ATPAD4/NA/ATHCHIB/RPP13/CAD1/RPS4/UGT71C3/NA/ATR4/PEP7/PROPEP6/CYP81F2/PH1/NDR1
#> GO:0045087                                                                                                                                                                                                                                                                                                                     FMO1/ALD1/AtGH3.12/SARD1/ATICS1/ATGST11/SGT1A/CBP60G/ATWRKY70/AZI1/ATL31/NA/PLA/SARD4/NA/NIMIN-2/NA/NHL3/CYP71A12/NA/BAL/AT-RLK3/SIB1/SOBER1/ATMES1/NA/CES101/ATHSPRO2/AtNUDT7/NSL1/RPM1/ATMYB72/ATOSM34/ATCHITIV/AT-SYR1/CYP79B2/ATSYP122/PCRK1/ALPHA-DOX1/NIMIN-1/CYP71A13/ATNPR3/ATMLO12/PROPEP3/PDI/AtBRN1/ATWRKY53/MEE54/TAO1/AGD2/ATL6/CYP71B15/MKS1/AtLYK4/ATMKK2/BAH1/ATCNGC2/ATSOT16/PROPEP2/BGLU26/ATPUB17/CHI/BIK1/AtNHR2B/RPP4/NA/AtCERK1/ATMEK1/AtTTM2/NA/ATNPR4/AtGILP/JAS1/ATPEPR1/ATHS83/ATWRKY33/CPR5/NA/GLIP1/ATPAD4/NA/ATHCHIB/RPP13/CAD1/RPS4/UGT71C3/NA/ATR4/PEP7/PROPEP6/CYP81F2/PH1/NDR1
#>            Count
#> GO:0009617   126
#> GO:0042493   119
#> GO:0042742   115
#> GO:0002376   103
#> GO:0006955    95
#> GO:0045087    93