Enriched ontologies for recognized organisms

This function returns the enriched biological processes in a set of genes, as compared to a background set of genes. The GO terms are retreived by the clusterProfiler package and the bioconductor organism databases (for example org.At.tair.db). The frequencies comparisons between GO terms in the genes of interest and the background are performed using Fischer tests, with 0.05 adjusted pvalues threshold. Those significantly enriched GO terms are then simplified with a semantic similarity threshold that can be chosen. It returns a dataframe containing, for each significant GO, their gene count, gene symbols, adjusted pvalue, description, etc.

enrich_go(
  genes,
  background,
  org = org.At.tair.db::org.At.tair.db,
  sim_cutoff = 0.85,
  GO_type = "BP",
  fdr = 0.05,
  pval = 0.05
)

Arguments

genes
gene set of interest. MUST be entrez IDs, that can be obtained from DIANE's functions convert_from_X(gene_IDs) depending on your organism. If your gene IDs are splicing aware, remove this information before with get_locus(gene_IDs).

background
gene set considered as the "universe" against which perfom the tests. MUST be entrez IDS that can be obtained from DIANE's functions convert_from_X(gene_IDs) depending on your organism. If your gene IDs are splicing aware, remove this information before with get_locus(gene_IDs).

org
organism, in the format of bioconductor organisms databases (e.g org.xx.xx.db). Possible values are, for DIANE's recognized organisms : org.Mm.eg.db, org.Hs.eg.db, org.Dm.eg.db, org.Ce.eg.db, org.EcK12.eg.db, org.At.tair.db

sim_cutoff
similarity cutoff to use for pooling similar GO terms

GO_type
character between BP (Biological Process), CC(Cellular Component) or MF (Molecular Function), depending on the GO subtype to use in the analysis

fdr
adjusted pvalues threshold, default id 0.05

pval
pvalues threshold, default id 0.05

Value

data.frame with gene significant gene ontologies, and their characteristics as columns

Examples

data("abiotic_stresses")
genes <- abiotic_stresses$heat_DEGs

genes <- get_locus(genes)
background <- get_locus(rownames(abiotic_stresses$normalized_counts))

genes <- convert_from_agi(genes)
background <- convert_from_agi(background)

go <- enrich_go(genes, background)
head(go)

#>                    ID                                  Description GeneRatio
#> GO:0009408 GO:0009408                             response to heat    50/499
#> GO:0006457 GO:0006457                              protein folding    38/499
#> GO:0006979 GO:0006979                 response to oxidative stress    31/499
#> GO:0009651 GO:0009651                      response to salt stress    31/499
#> GO:0071456 GO:0071456                 cellular response to hypoxia    25/499
#> GO:0036294 GO:0036294 cellular response to decreased oxygen levels    25/499
#>              BgRatio       pvalue     p.adjust       qvalue
#> GO:0009408 230/15734 7.272616e-28 8.923500e-25 8.811348e-25
#> GO:0006457 174/15734 1.761225e-21 1.080512e-18 1.066932e-18
#> GO:0006979 417/15734 1.041411e-05 5.555703e-04 5.485878e-04
#> GO:0009651 443/15734 3.378788e-05 1.727405e-03 1.705695e-03
#> GO:0071456 224/15734 4.885258e-08 4.995176e-06 4.932397e-06
#> GO:0036294 227/15734 6.356699e-08 5.571193e-06 5.501173e-06
#>                                                                                                                                                                                                                                                                                                                                                                                             geneID
#> GO:0009408 Hop3/ATFKBP65/HSP18.5/HSP17.6B/ATHSFA2/ATHS83/ATHSP101/ATMBF1C/DjC53/ATHSP17.4/AT-HSP17.6A/HSP70T-2/ATBAG6/HSP17.6A/HSP17.8/HKL/ATFKBP62/CLPB-M/ATERDJ3A/ATHSP70/HSP17.6II/HSP17.6C/CPN10/AtHsp90.5/HSP60/HSC70-5/HSP17.4B/AT-P5C1/HSP15.7/SES1/Fes1A/AtHsp90.2/HSA32/BIP1/BIP/AR192/DREB2/AtGolS1/ATHSP22.0/PUB48/HSP21/HSP18.2/Hsp70b/AT-HSFA3/HTT1/DREB2/AtPARK13/AtERF53/ERF54/BIP3
#> GO:0006457                                                                                           ATFKBP65/HSP18.5/HSP17.6B/NA/ATHS83/ATHSP17.4/AT-HSP17.6A/HSP70T-2/ATBAG6/HSP17.6A/HSP17.8/ATFKBP62/DNAJ/CLPB-M/NA/ATHSP70/HSP17.6II/HSP17.6C/CPN10/AtHsp90.5/HSP60/HSC70-5/AtHsp90-6/HSP17.4B/HSP15.7/SES1/AtHsp90.2/ATCNX1/BIP1/BIP/AR192/Cpn60beta2/CYP40/ATHSP22.0/NA/HSP18.2/Hsp70b/BIP3
#> GO:0006979                                                                                                                                                             HSP18.5/HSP17.6B/ATHSFA2/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HSP17.6II/HSP17.6C/HSP17.4B/HSP15.7/NA/DREB2/AtGolS1/ATHSP22.0/NA/APX1B/ATWRKY54/AtGolS3/HSP18.2/NA/NA/ANAC092/PRX40/ATFER3/BIP3/PRX9/ATWRKY30/ATFER1/NA/NA
#> GO:0009651                                                                                                                                    HSP18.5/HSP17.6B/ATGRP7/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HSP17.6II/HSP17.6C/AtHsp90.5/HSP17.4B/AT-P5C1/HSP15.7/Fes1A/AtHsp90.2/ABX45/AtGolS1/ATHSP22.0/HSP18.2/CFM9/ANAC092/AtLtpI-11/5PTASE11/AtERF53/AtLEA3/FAR5/AtRLP30/NA/ERD7/WSD1/ATTSPO
#> GO:0071456                                                                                                                                                                                                                  ATHSFA2/ATHSP101/ATMBF1C/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HKL/HSP17.6II/AT-HSFB2B/HSP17.4B/HSP15.7/ETR2/DREB2/NA/MPSR1/ATHSP22.0/NA/STMP5/NA/NA/CYP81F2/NA/NA/NA
#> GO:0036294                                                                                                                                                                                                                  ATHSFA2/ATHSP101/ATMBF1C/ATHSP17.4/AT-HSP17.6A/HSP17.6A/HSP17.8/HKL/HSP17.6II/AT-HSFB2B/HSP17.4B/HSP15.7/ETR2/DREB2/NA/MPSR1/ATHSP22.0/NA/STMP5/NA/NA/CYP81F2/NA/NA/NA
#>            Count
#> GO:0009408    50
#> GO:0006457    38
#> GO:0006979    31
#> GO:0009651    31
#> GO:0071456    25
#> GO:0036294    25