estimateDispersion

This function computes the common, trended, and tag-wise dispersions of the gene expression dataset. Once, the dispersion is estimated, it fits a negative binomial model for each gene using default design : ~group+0. In this configuration, the log average expression of each gene is approximated by a linear combination of each of the conditions. It returns a the value of a glmFit object from the package edgeR, containing the attributes of the tcc object given, plus fitted values and GLM coefficients, among other indicators of the fitting procedure.

estimateDispersion(tcc, conditions = NULL)

Arguments

tcc
TCC-class object countaining counts, and norm factors, such as obtained after DIANE::normalize() and DIANE::filter_low_counts()

conditions
if NULL (default), takes the split before the character '_' as condition names. else, the conditions specified by the user, as a character vector. Its order should match the columns names of the expression matrix used to build the tcc object.

Value

glmFit

# Note

The return value is meant to be used directly as the parameter of DIANE::estimateDEGs()

Examples

data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)

#> TCC::INFO: Calculating normalization factors using tmm ...

#> TCC::INFO: Done.

threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

#>  [1] "cNF" "cNF" "cNF" "cnF" "cnF" "cnF" "cNf" "cnf" "cnf" "cNf" "cNf" "cnf"
#>       cnf cnF cNf cNF
#> cNF_3   0   0   0   1
#> cNF_2   0   0   0   1
#> cNF_1   0   0   0   1
#> cnF_2   0   1   0   0
#> cnF_1   0   1   0   0
#> cnF_3   0   1   0   0
#> cNf_1   0   0   1   0
#> cnf_2   1   0   0   0
#> cnf_1   1   0   0   0
#> cNf_2   0   0   1   0
#> cNf_3   0   0   1   0
#> cnf_3   1   0   0   0
#> attr(,"assign")
#> [1] 1 1 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$groups
#> [1] "contr.treatment"
#> 

#> Warning: norm factors don't multiply to 1