Fits a Poisson generalized linear model to a set of genes

The idea is to extract the importance and effect of each factor. To do so, the expression of each gene is modelled as a Poisson distribution. The log of its parameter (the expected value) is approximated by a linear combination of the factors in the experiment. The coefficients associated to each factors are estimated to fit gene expression, and can be insightful to characterize genes behavior in a particular cluster. The model with interactions is considered. It your design in not a complete corssed design, the interaction term will be null.

fit_glm(normalized_counts, genes, design, factors = colnames(design))

Arguments

normalized_counts
normalized counts

genes
genes belonging to a specific expression-based clusters

design
experimental design as a dataframe

factors
factors to use for the fit (defalut is all the factors of the design)

Value

glm object

Note

Note that we can only apply a glm fit to a set of genes that have very close expression profiles accros conditions, else we would have to introduce a new variable related to the genes themselves.

Examples

library(DIANE)
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)

#> TCC::INFO: Calculating normalization factors using tmm ...

#> TCC::INFO: Done.

threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

#> Warning: norm factors don't multiply to 1

topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes
clustering <- DIANE::run_coseq(conds = unique(demo_data_At$conditions), data = normalized_counts, genes = genes, K = 6:9)

#> ****************************************
#> coseq analysis: Poisson approach & none transformation
#> K = 6 to 9 
#> Use set.seed() prior to running coseq for reproducible results.
#> ****************************************
#> Running K = 6 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 14.1847548994142"
#> [1] "Log-like diff: 10.0310272898039"
#> [1] "Log-like diff: 10.298485186429"
#> [1] "Log-like diff: 13.992219216283"
#> [1] "Log-like diff: 17.1009464677972"

#> Running K = 7 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 642.489088683807"
#> [1] "Log-like diff: 73.6556548560898"
#> [1] "Log-like diff: 36.9400212284619"
#> [1] "Log-like diff: 86.2699015610276"
#> [1] "Log-like diff: 47.3096994450798"

#> Running K = 8 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 157.229844735799"
#> [1] "Log-like diff: 444.25478246178"
#> [1] "Log-like diff: 109.846081027161"
#> [1] "Log-like diff: 6.18835455160118"
#> [1] "Log-like diff: 45.3760373939516"

#> Running K = 9 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 187.508275240089"
#> [1] "Log-like diff: 278.457512046144"
#> [1] "Log-like diff: 121.525492708007"
#> [1] "Log-like diff: 184.259286907845"
#> [1] "Log-like diff: 5.94173633496066"

#> Note: slope heuristics for model selection only applied if > 10 models are fit.

genes_cluster <- DIANE::get_genes_in_cluster(clustering$membership, cluster = 3)
glm <- DIANE::fit_glm(normalized_counts, genes_cluster, demo_data_At$design)
summary(glm)

#> 
#> Call:
#> glm(formula = formula, family = poisson(link = "log"), data = glmData)
#> 
#> Deviance Residuals: 
#>    Min      1Q  Median      3Q     Max  
#> -68.76  -39.65  -27.40   -7.55  504.90  
#> 
#> Coefficients:
#>                Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)   7.7974443  0.0008358 9329.01   <2e-16 ***
#> Nitrate      -0.5480407  0.0013810 -396.85   <2e-16 ***
#> Iron         -1.3511415  0.0018429 -733.16   <2e-16 ***
#> Nitrate:Iron  0.1199545  0.0029569   40.57   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> (Dispersion parameter for poisson family taken to be 1)
#> 
#>     Null deviance: 8239249  on 2351  degrees of freedom
#> Residual deviance: 7043884  on 2348  degrees of freedom
#> AIC: 7060286
#> 
#> Number of Fisher Scoring iterations: 6
#>