Fits a Poisson generalized linear model to a set of genes

The idea is to extract the importance and effect of each factor. To do so, the expression of each gene is modelled as a Poisson distribution. The log of its parameter (the expected value) is approximated by a linear combination of the factors in the experiment. The coefficients associated to each factors are estimated to fit gene expression, and can be insightful to characterize genes behavior in a particular cluster. The model with interactions is considered. It your design in not a complete corssed design, the interaction term will be null.

fit_glm(normalized_counts, genes, design, factors = colnames(design))

Arguments

normalized_counts
normalized counts

genes
genes belonging to a specific expression-based clusters

design
experimental design as a dataframe

factors
factors to use for the fit (defalut is all the factors of the design)

Value

glm object

Note

Note that we can only apply a glm fit to a set of genes that have very close expression profiles accros conditions, else we would have to introduce a new variable related to the genes themselves.

Examples

library(DIANE)
data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)

#> TCC::INFO: Calculating normalization factors using tmm ...

#> TCC::INFO: Done.

threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

#> Warning: norm factors don't multiply to 1

topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes
clustering <- DIANE::run_coseq(conds = unique(demo_data_At$conditions),
data = normalized_counts, genes = genes, K = 6:9)

#> ****************************************
#> coseq analysis: Poisson approach & none transformation
#> K = 6 to 9 
#> Use set.seed() prior to running coseq for reproducible results.
#> ****************************************
#> Running K = 6 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 14.1847548994142"
#> [1] "Log-like diff: 10.0310272898039"
#> [1] "Log-like diff: 10.298485186429"
#> [1] "Log-like diff: 13.992219216283"
#> [1] "Log-like diff: 17.1009464677972"

#> Running K = 7 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 0.278067337668776"
#> [1] "Log-like diff: 0.0469732341073055"
#> [1] "Log-like diff: 0.0160307238038966"
#> [1] "Log-like diff: 0.00449461122679562"
#> [1] "Log-like diff: 0.00135304200780872"

#> Running K = 8 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 146.370285499767"
#> [1] "Log-like diff: 1.19442753947416"
#> [1] "Log-like diff: 0.581354155369221"
#> [1] "Log-like diff: 0.358276221962332"
#> [1] "Log-like diff: 0.233443789817207"

#> Running K = 9 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 3.28789536868446"
#> [1] "Log-like diff: 0.930758150992422"
#> [1] "Log-like diff: 1.07079815105465"
#> [1] "Log-like diff: 0.922957424227423"
#> [1] "Log-like diff: 0.579147871833758"

#> Note: slope heuristics for model selection only applied if > 10 models are fit.

genes_cluster <- DIANE::get_genes_in_cluster(clustering$membership, cluster = 3)
glm <- DIANE::fit_glm(normalized_counts, genes_cluster, demo_data_At$design)
summary(glm)

#> 
#> Call:
#> glm(formula = formula, family = poisson(link = "log"), data = glmData)
#> 
#> Deviance Residuals: 
#>    Min      1Q  Median      3Q     Max  
#> -68.58  -39.14  -27.35   -7.64  505.47  
#> 
#> Coefficients:
#>                Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)   7.7921932  0.0008338 9345.60   <2e-16 ***
#> Nitrate      -0.5478611  0.0013775 -397.72   <2e-16 ***
#> Iron         -1.3496343  0.0018373 -734.58   <2e-16 ***
#> Nitrate:Iron  0.1225181  0.0029460   41.59   <2e-16 ***
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> (Dispersion parameter for poisson family taken to be 1)
#> 
#>     Null deviance: 8251186  on 2375  degrees of freedom
#> Residual deviance: 7052896  on 2372  degrees of freedom
#> AIC: 7069494
#> 
#> Number of Fisher Scoring iterations: 6
#>