run_coseq

run_coseq

run_coseq(conds, genes, data, K = 6:12)

Arguments

conds
Condition names to be used for clustering

genes
Genes used as an input for the clustering

data
normalized counts (MUST be normalized!)

K
number of clusters range

Value

named list containing the coseq run result as "model", and the cluster membership for each gene as "membership"

Examples

data("demo_data_At")
tcc_object <- DIANE::normalize(demo_data_At$raw_counts, demo_data_At$conditions, iteration = FALSE)

#> TCC::INFO: Calculating normalization factors using tmm ...

#> TCC::INFO: Done.

threshold = 10*length(demo_data_At$conditions)
tcc_object <- DIANE::filter_low_counts(tcc_object, threshold)
normalized_counts <- TCC::getNormalizedData(tcc_object)
fit <- DIANE::estimateDispersion(tcc = tcc_object, conditions = demo_data_At$conditions)

#>  [1] "cNF" "cNF" "cNF" "cnF" "cnF" "cnF" "cNf" "cnf" "cnf" "cNf" "cNf" "cnf"
#>       cnf cnF cNf cNF
#> cNF_3   0   0   0   1
#> cNF_2   0   0   0   1
#> cNF_1   0   0   0   1
#> cnF_2   0   1   0   0
#> cnF_1   0   1   0   0
#> cnF_3   0   1   0   0
#> cNf_1   0   0   1   0
#> cnf_2   1   0   0   0
#> cnf_1   1   0   0   0
#> cNf_2   0   0   1   0
#> cNf_3   0   0   1   0
#> cnf_3   1   0   0   0
#> attr(,"assign")
#> [1] 1 1 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$groups
#> [1] "contr.treatment"
#> 

#> Warning: norm factors don't multiply to 1

topTags <- DIANE::estimateDEGs(fit, reference = "cNF", perturbation = "cnF", p.value = 0.01)
genes <- topTags$table$genes
clustering <- DIANE::run_coseq(conds = unique(demo_data_At$conditions), data = normalized_counts, genes = genes, K = 6:9)

#> ****************************************
#> coseq analysis: Poisson approach & none transformation
#> K = 6 to 9 
#> Use set.seed() prior to running coseq for reproducible results.
#> ****************************************
#> Running K = 6 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 9.83236934009439"
#> [1] "Log-like diff: 14.169959903156"
#> [1] "Log-like diff: 10.0395549234466"
#> [1] "Log-like diff: 10.3368664861309"
#> [1] "Log-like diff: 14.0524304044068"

#> Running K = 7 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 0.277257664246275"
#> [1] "Log-like diff: 0.0468137651419909"
#> [1] "Log-like diff: 0.0159963850967237"
#> [1] "Log-like diff: 0.00448288231594685"
#> [1] "Log-like diff: 0.00135403044763116"

#> Running K = 8 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 1133.43946119693"
#> [1] "Log-like diff: 50.1545745216742"
#> [1] "Log-like diff: 830.803607148806"
#> [1] "Log-like diff: 551.200343350271"
#> [1] "Log-like diff: 736.850624551559"

#> Running K = 9 ...
#> [1] "Initialization: 1"
#> [1] "Log-like diff: 174.321196427855"
#> [1] "Log-like diff: 136.811287722964"
#> [1] "Log-like diff: 302.763742413295"
#> [1] "Log-like diff: 108.530814716716"
#> [1] "Log-like diff: 117.160605134435"

#> Note: slope heuristics for model selection only applied if > 10 models are fit.